src/rapgeneratorservice/requirements.txt
confluent-kafka
glob
fnmatch
google.cloud
aubio
numpy
pydub
scipy
six
pythonjsonlogger